permeate uint for Ids throughout app
not sure I'll stick with this long-term yet, but works for now.

ref https://github.com/PomeloFoundation/Pomelo.EntityFrameworkCore.MySql/issues/601#issuecomment-396346202

+++ src/FilterLists.Api/V1/Controllers/ListsController.cs
@@ -21,7 +21,8 @@ namespace FilterLists.Api.V1.Controllers
 
         [HttpGet]
         [Route("{id}")]
-        public async Task<IActionResult> GetById(int id) => Json(await _filterListService.GetDetailsAsync(id));
+        //TODO: respond with appropriate exception if negative id queried
+        public async Task<IActionResult> GetById(int id) => Json(await _filterListService.GetDetailsAsync((uint)id));
 
         [HttpGet("seed")]
         public async Task<IActionResult> Seed() => Json(await SeedService.GetAllAsync<FilterList, FilterListSeedDto>());
